Update for anki 23.10
anki 23.10 includes protocol changes (see linked PR, and others) so we
need to update the anki base commit and patch.

I've aligned the rusqlite dependency to fix rust build, and while I was
it aligned a few more where rust would be building an extra version of
the lib for anki.

Link: https://github.com/ankitects/anki/pull/2647